refactor:#54 훅 반환값 기능단위 분리
